requirements.txt
fastapi
pydantic
pydantic[email]
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
PyJWT
passlib
argon2_cffi
pytest
pytest-asyncio